requirements.txt
# ── LLM & Agent Framework ─────────────────
langchain>=0.3.0
langchain-openai>=0.3.0
langchain-community>=0.3.0
langgraph>=0.2.0

# ── Embedding & Vector Store ──────────────
faiss-cpu>=1.9.0
sentence-transformers>=3.0.0
langchain-huggingface>=0.1.0

# ── Document Loading ──────────────────────
pymupdf>=1.25.0

# ── Utilities ─────────────────────────────
python-dotenv>=1.0.0
